# apps/be/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BE API",
    "version": "1.0.0",
    "description": "Backend API - Auto-generated from JSDoc"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development"
    }
  ],
  "paths": {
...
    "/api/cards/{id}": {
      "get": {
        "tags": [
          "Cards"
        ],
        "summary": "Get card by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Card found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "stackId": {
                      "type": "integer"
                    },
                    "name": {
                      "type": "string"
                    },
                    "cover": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Card not found"
          }
        }
      },
      "put": {
        "tags": [
          "Cards"
        ],
        "summary": "Update a card",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "cover": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "stackId": {
                    "type": "integer",
                    "description": "Move card to a different stack"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Card updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "stackId": {
                      "type": "integer"
                    },
                    "name": {
                      "type": "string"
                    },
                    "cover": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Card not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Cards"
        ],
        "summary": "Delete a card",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Card deleted"
          },
          "404": {
            "description": "Card not found"
          }
        }
      }
    },
    "/api/stacks/{stackId}/cards": {
      "get": {
        "tags": [
          "Cards"
        ],
        "summary": "Get all cards in a stack",
        "parameters": [
          {
            "name": "stackId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of cards",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "stackId": {
                        "type": "integer"
                      },
                      "name": {
                        "type": "string"
                      },
                      "cover": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string"
                      },
                      "createdAt": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Stack not found"
          }
        }
      },
      "post": {
        "tags": [
          "Cards"
        ],
        "summary": "Create a card in a stack",
        "parameters": [
          {
            "name": "stackId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "cover"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "cover": {
                    "type": "string",
                    "description": "Image URL for the card cover"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Card created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "stackId": {
                      "type": "integer"
                    },
                    "name": {
                      "type": "string"
                    },
                    "cover": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Stack not found"
          }
        }
      }
    },
    "/api/cards/{id}/move": {
      "patch": {
        "tags": [
          "Cards"
        ],
        "summary": "Move a card to a different stack",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "stackId"
                ],
                "properties": {
                  "stackId": {
                    "type": "integer",
                    "description": "Target stack ID"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Card moved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "stackId": {
                      "type": "integer"
                    },
                    "name": {
                      "type": "string"
                    },
                    "cover": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Card or stack not found"
          }
        }
      }
    },
...
  },
...
}